deltagolomb: Redesigned interface to mirror that of pkg/compress.
In particular, there's now a NewExpGolombEncoder and NewExpGolombDecoder
function that take an io.Writer and io.Reader, respectively.  The
encoder and decoder now operate synchronously, triggered by calls to
Read and Write, instead of running in their own goroutines and
reading from channels.  State management is more complex, but the
code is quite a bit faster (~4x).
Thanks to roger peppe on golang-nuts for prompting the evaluation of
the new interface.