Optimized the encoding path to operate byte-at-a-time instead of bit-at-a-time.
Replaced the function addBit(bit) with addBits(uint, nbits) and a specialized
addZeros(nbits) that operate on bytes or partial bytes at a time instead of
bits at a time.  Speeds encoding up by a further 50%, now running at about
193-197 ns per write of 5 ints on my macbook pro (~25M symbols/second).
Added an optimized path for +1 and -1.

Corrected comment for Add function and made it non-exported in favor of
a new WriteInt() function analogous to Write().